requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.42.0
blis==0.7.11
captum==0.7.0
catalogue==2.0.10
#causal-conv1d==1.2.0.post2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.1.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
datasets==2.19.2
dill==0.3.8
docker-pycreds==0.4.0
einconv==0.1.0
einops==0.8.0
evaluate==0.4.3
faiss-cpu==1.8.0
ferret-xai==0.4.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.6
imageio==2.36.1
importlib_resources==6.1.2
inquirerpy==0.3.4
inseq @ git+https://github.com/inseq-team/inseq.git@a7f77fd86e1841763596ac7743dc4fb9f0ca60dc
jaxtyping==0.2.27
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
kiwisolver==1.4.5
kronfluence==1.0.1
langcodes==3.3.0
lazy_loader==0.4
lightgbm==4.3.0
lime==0.2.0.1
llvmlite==0.43.0
#mamba-ssm==1.2.0.post1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mauve-text==0.3.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
nmslib==2.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.1
openai==0.27.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optimum==1.24.0
packaging==23.2
pandas==2.2.3
peft==0.13.2
pfzy==0.3.4
pillow==10.2.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.50
protobuf==3.20.2
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.6.1
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pyjnius==1.6.1
pyparsing==3.1.2
pyserini==0.21.0
python-dateutil==2.8.2
pytreebank==0.2.7
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.3
rich==13.7.1
rouge-score==0.1.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
shap==0.46.0
six==1.16.0
slicer==0.0.8
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.8.30
tokenizers==0.19.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.0
transformers==4.43.1
triton==2.2.0
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.18.7
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0